+++ Changes.md
@@ -3,6 +3,8 @@ History
 
 Version 403.1.1 - TBR
 ----------------------------
+1. Fix 'Import a format_grid course into a format_topics course fails and produces "undefined method"' - #193,
+   thanks to @Syxton(Matt Davidson).
 
 Version 403.1.0 - 15/10/2023
 ----------------------------