+++ Changes.md
@@ -14,7 +14,7 @@ Version 402.2.1b - 14/10/2023
 -----------------------------
 1. Fix 'Webp image causes an error and makes course unaccessible' - #190.
 2. Fix "Hidden sections are shown as not available" doesn't seem to work. - #192.
-3. Add 'Flex containers' justify-content property options' - #191.  Sponsored by AAU.
+3. Add 'Flex containers' justify-content property options' - #191.
 4. Fix 'Problems with creating multiple courses from template via CSV in Grid Format' - #189.
 5. Make 'Main content page' link on single section page more accessible.
 6. Fix 'Mouseover for shortened titles' - #187.