+++ Changes.md
@@ -1,7 +1,7 @@
 History
 =============
 
-Version 403.1.1 - TBR
+Version 403.1.1 - 22/01/2024
 ----------------------------
 1. Fix 'Import a format_grid course into a format_topics course fails and produces "undefined method"' - #193,
    thanks to @Syxton(Matt Davidson).